+++ app/controllers/stocks_controller.rb
@@ -6,8 +6,9 @@ class StocksController < ApplicationController
     @stock = Stock.friendly.find(params[:id])
     @new_stock = Stock.new
     @news_hash = news(@stock, @enddate)
-    @basic_info = @stock.basic_info
+    @basic_info = @stock.yahooapi
     @reddit_articles = RedditSearch.call(@stock.ticker)
+
   end
 
   # /stocks(.:format)
