+++ app/views/shared/_confirm_modals.html.erb
@@ -14,4 +14,4 @@
   </form>
 </dialog>
 
-<%# Logic is in app/javascript/application.js %>
+<%# Logic is in app/javascript/controllers/application.js %>
