+++ app/views/shared/_stocks-container.html.erb
@@ -9,9 +9,9 @@
               <th scope="col" class="px-3 py-3.5 text-left text-sm font-semibold text-gray-900">Name</th>
               <th scope="col" class="px-3 py-3.5 text-left text-sm font-semibold text-gray-900">Ticker</th>
               <th scope="col" class="px-3 py-3.5 text-left text-sm font-semibold text-gray-900">Last</th>
-              <th scope="col" class="px-3 py-3.5 text-left text-sm font-semibold text-gray-900">High</th>
-              <th scope="col" class="px-3 py-3.5 text-left text-sm font-semibold text-gray-900">Low</th>
-              <th scope="col" class="px-3 py-3.5 text-left text-sm font-semibold text-gray-900">Change</th>
+              <th scope="col" class="px-3 py-3.5 text-left text-sm font-semibold text-gray-900">Daily High</th>
+              <th scope="col" class="px-3 py-3.5 text-left text-sm font-semibold text-gray-900">Daily Low</th>
+              <th scope="col" class="px-3 py-3.5 text-left text-sm font-semibold text-gray-900">Change $</th>
               <th scope="col" class="px-3 py-3.5 text-left text-sm font-semibold text-gray-900">Change %</th>
               <th class="px-3 py-3.5 text-left text-sm font-semibold text-gray-900"></th>
             </tr>
@@ -22,14 +22,20 @@
               <% stocks.each_with_index do |stock, index| %>
                 <tr class="group hover:bg-indigo-500 text-white cursor-pointer" onclick="window.location='<%= stock_url(stock) %>'">
                   <td class="whitespace-nowrap py-4 pl-4 pr-3 text-sm font-medium text-gray-900 sm:pl-6 group-hover:text-white"> <%= index + 1 %></td>
-                  <td class="whitespace-nowrap px-3 py-4 text-sm text-gray-500 group-hover:text-white"><%= stock.name %></td>
+                  <td class="whitespace-nowrap px-3 py-4 text-sm text-gray-500 group-hover:text-white"><%= stock.yahooapi["shortName"] %></td>
                   <td class="whitespace-nowrap px-3 py-4 text-sm text-gray-500 group-hover:text-white"><%= stock.ticker %></td>
                   <%# Interpolate data from API to Last, High & Low Below %>
-                  <td class="whitespace-nowrap px-3 py-4 text-sm text-gray-500 group-hover:text-white/50">$55.00</td>
-                  <td class="whitespace-nowrap px-3 py-4 text-sm text-gray-500 group-hover:text-white/50">$62.00</td>
-                  <td class="whitespace-nowrap px-3 py-4 text-sm text-gray-500 group-hover:text-white/50">$52.00</td>
-                  <td class="whitespace-nowrap px-3 py-4 text-sm text-gray-500 group-hover:text-white/50">+ $3.00</td> <%# Last vs Last %>
-                  <td class="whitespace-nowrap px-3 py-4 text-sm text-gray-500 group-hover:text-white/50">+1.02%</td> <%# Last vs Last as %%>
+                  <td class="whitespace-nowrap px-3 py-4 text-sm text-gray-500 group-hover:text-white/50">$<%= stock.yahooapi["regularMarketPrice"].round(2) %></td>
+                  <td class="whitespace-nowrap px-3 py-4 text-sm text-gray-500 group-hover:text-white/50">$<%= stock.yahooapi["regularMarketDayHigh"].round(2) %></td>
+                  <td class="whitespace-nowrap px-3 py-4 text-sm text-gray-500 group-hover:text-white/50">$<%= stock.yahooapi["regularMarketDayLow"].round(2) %></td>
+                  <%# Display Green or Red depending on whether Up or Down %>
+                  <% if stock.yahooapi["regularMarketChange"].negative? %>
+                    <td class="whitespace-nowrap px-3 py-4 text-sm text-emerald-500 group-hover:text-white/50">$<%= stock.yahooapi["regularMarketChange"].round(2).abs %></td>
+                    <td class="whitespace-nowrap px-3 py-4 text-sm text-emerald-500 group-hover:text-white/50"><%= stock.yahooapi["postMarketChangePercent"].round(2).abs %>%</td>
+                  <% else %>
+                    <td class="whitespace-nowrap px-3 py-4 text-sm text-rose-500 group-hover:text-white/50">$<%= stock.yahooapi["regularMarketChange"].round(2).abs %></td>
+                    <td class="whitespace-nowrap px-3 py-4 text-sm text-rose-500 group-hover:text-white/50"><%= stock.yahooapi["postMarketChangePercent"].round(2).abs %>%</td>
+                  <% end %>
                   <%# Only Display Delete Option if Portfolio is present %>
                   <% if portfolio %>
                     <td class="whitespace-nowrap px-3 py-4 text-sm text-gray-500 table-cell">
