+++ config/initializers/algoliasearch.rb
@@ -1,0 +1,4 @@
+AlgoliaSearch.configuration = {
+  application_id: 'K0X97I2FU8',
+  api_key: ENV['ALGOLIA_ADMIN_ID']
+}
